Index Buffers(.cpp)/Shader file(.shader)
Used index buffer to eliminate the repetition of data due to multiple points being the same as each resulting. Learned that this return is a waste of memory and can be solved by doing the following in the code. Linked .shader file through ifstream function passing thru the file-path as a parameter.